requirements.txt
-i https://pypi.org/simple
boto3==1.17.84
botocore==1.20.84
click==8.0.1
flask-sqlalchemy==2.5.1
flask==2.0.1
greenlet==1.1.0 ; python_version >= '3'
importlib-metadata==4.3.1 ; python_version < '3.8'
itsdangerous==2.0.1
jinja2==3.0.1
jmespath==0.10.0
markupsafe==2.0.1
pyjwt==2.1.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
sqlalchemy==1.4.17
typing-extensions==3.10.0.0 ; python_version < '3.8'
urllib3==1.26.5
werkzeug==2.0.1
zipp==3.4.1
